Add ATF submodule and the patch to add FIQ controller service.
In order to allow the guests direct access to the gic controller, Bao
has to forward all the IRQs it receives to EL1/0; this means that the
only way for it to receive interrupts is to use FIQs, that are only
controllable in secure mode.

This patch implements a simple FIQ controller in the ATF tha Bao can
use to configure secure interrupts, and adds instructions on how to use
a custom bl31 on zynqmp.