arm, interrupts: use IPI 15 instead of 1
The linux kernel uses the SGIs 0 to 9 in gicv2 for internal purposes,
and recommends kernel developers not to use other SGIs.

When the gic virtualization interface is used there is no problem of
conflict when using the same SGIs, but in a direct gic controller
access scenario a guest won't be allowed to register an SGI that has
the same ID of any of those used by Bao.

This commit changes the IPI used by Bao to avoid conflict with the Linux
Kernel.