arm, gicv2: allow guests to directly access the controller.
In gicv2, the usual way of allowing a guest to use interrupts is to
use the virtual gic interface [1]; this, however introduces a
measurable amount of latency to every interrupt, especially when there
is interference of other guests.

Being Bao purely a partitioning Hypervisor, it doesn't actually need to
do much work over the interrupts, and it could in fact just act as a
mediator for the guests to access the gic distributor, while letting
them use the physical gic controller, skipping the gic virtual
interface altogether and thus completely removing overhead for guests
interrupts.

This patch allows the guests to access the controller and removes all
use of the virual gic interfaces, though leaving in place the access
emulation to the gic distributor.

Bao now directly forwards the interrupts it receives to EL1/0 and uses
secure interupts for its internal needs.

[1] https://static.docs.arm.com/ihi0069/d/IHI0069D_gic_architecture_specification.pdf#G8.7106169